Safari-611.1.10/LayoutTests/fast/dom/parsed-document-dispatchEvent-expected.txt

Test that it is possible to dispatchEvent() works on documents created using DOMParser.parseFromString().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS wasEventDispatched is false
PASS wasEventDispatched is true
PASS successfullyParsed is true

TEST COMPLETE
